Explicitly cast id to uint to match ...Job constructors - avoids crash on import due to 'Constructor on ... not found'

+++ src/Clients/MainApp/FSpot/JobStore.cs
@@ -156,7 +156,7 @@ namespace FSpot {
     						DateTimeUtil.FromDateTime (run_at),
     						Convert.ToInt32 (job_priority)));
     
-                    Job job = (Job)Activator.CreateInstance (job_type, id, job_options, run_at, job_priority, true);
+			Job job = (Job)Activator.CreateInstance (job_type, (uint) id, job_options, run_at, job_priority, true);
     
     		AddToCache (job);
     		job.Finished += HandleRemoveJob;
